refactor: LLM providers logic to dispatcher module
