+++ lua/gp/config.lua
@@ -517,12 +517,6 @@ local config = {
 			local copy = vim.deepcopy(plugin)
 			local key = copy.config.openai_api_key or ""
 			copy.config.openai_api_key = key:sub(1, 3) .. string.rep("*", #key - 6) .. key:sub(-3)
-			for provider, _ in pairs(copy.providers) do
-				local s = copy.providers[provider].secret
-				if s and type(s) == "string" then
-					copy.providers[provider].secret = s:sub(1, 3) .. string.rep("*", #s - 6) .. s:sub(-3)
-				end
-			end
 			local plugin_info = string.format("Plugin structure:\n%s", vim.inspect(copy))
 			local params_info = string.format("Command params:\n%s", vim.inspect(params))
 			local lines = vim.split(plugin_info .. "\n" .. params_info, "\n")
